coap-engine,coap-transactions: Fix handling of separate response
A response from a server might come as a "separate response" where
first an empty message ACK is sent to tell the client there is no need
to retransmit, and that it should expect a response message soon. (RFC 7252 §5.2.2)

To stop the client from retransmitting we stop the retransmit timer of
the transaction when we receive the empty message ACK.

The response message is sent with a new Message ID chosen by the
server, and if its type is CON it should be ACK:ed by the client to
inform the server that the response has been successfully received.

To be able to handle the above we also need to fix how transactions
are managed and matched. According to RFC, request/response should be
matched using the token and endpoint information only. (RFC 7252 §5.3)