Fix prop mode receiving_packet under CSMA
The current version of this function relies on the value of the MDMSOFT interrupt flag. The radio sets this flag when it detects a sync word, but it does _not_ clear the flag automatically.

The current version of the function assumes that the caller will keep calling the function repetitively until frame RX has finished, at which point we can clear the flag. This assumption is correct for TSCH.

This assumption however is _not_ correct for CSMA, which will sometimes make a one-off call to this function here. In these cases, the logic of how/when to clear MDMSOFT will fail and the function will very often erroneously signal ongoing frame reception.